+++ meta/NotificationPortStateChange.cpp
@@ -2,7 +2,7 @@
 
 #include "swss/logger.h"
 
-#include "meta/sai_serialize.h"
+#include "sai_serialize.h"
 
 using namespace sairedis;
