+++ vslib/src/SwitchStateBaseFdb.cpp
@@ -5,7 +5,7 @@
 #include "swss/select.h"
 
 #include "meta/sai_serialize.h"
-#include "lib/inc/NotificationFdbEvent.h"
+#include "meta/NotificationFdbEvent.h"
 
 #include <linux/if_ether.h>
 #include <arpa/inet.h>
